update: livewire sortBy to trait and add parameter type
