+++ app/Http/Livewire/AnnounceSearch.php
@@ -14,6 +14,7 @@
 namespace App\Http\Livewire;
 
 use App\Models\Announce;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -22,6 +23,7 @@ use Livewire\WithPagination;
  */
 class AnnounceSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public int $perPage = 50;
@@ -71,17 +73,6 @@ class AnnounceSearch extends Component
             ->when($this->userId !== '', fn ($query) => $query->where('user_id', '=', $this->userId))
             ->when($this->sortField !== '', fn ($query) => $query->orderBy($this->sortField, $this->sortDirection))
             ->paginate($this->perPage);
-    }
-
-    final public function sortBy(string $field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
