+++ app/Http/Livewire/ApikeySearch.php
@@ -15,6 +15,7 @@ namespace App\Http\Livewire;
 
 use App\Models\Apikey;
 use App\Models\User;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -23,6 +24,7 @@ use Livewire\WithPagination;
  */
 class ApikeySearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public string $username = '';
@@ -69,16 +71,5 @@ class ApikeySearch extends Component
         return view('livewire.apikey-search', [
             'apikeys' => $this->apikeys,
         ]);
-    }
-
-    final public function sortBy(string $field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 }
