+++ app/Http/Livewire/FailedLoginSearch.php
@@ -14,6 +14,7 @@
 namespace App\Http\Livewire;
 
 use App\Models\FailedLoginAttempt;
+use App\Traits\LivewireSort;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
@@ -21,6 +22,7 @@ use Livewire\WithPagination;
 
 class FailedLoginSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public string $username = '';
@@ -78,17 +80,6 @@ class FailedLoginSearch extends Component
             ->when($this->ipAddress, fn ($query) => $query->where('ip_address', 'LIKE', $this->ipAddress.'%'))
             ->orderBy($this->sortField, $this->sortDirection)
             ->paginate($this->perPage);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
