+++ app/Http/Livewire/InviteLogSearch.php
@@ -15,12 +15,14 @@ namespace App\Http\Livewire;
 
 use App\Models\Invite;
 use App\Models\User;
+use App\Traits\LivewireSort;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class InviteLogSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public string $sender = '';
@@ -142,16 +144,5 @@ class InviteLogSearch extends Component
         return view('livewire.invite-log-search', [
             'invites' => $this->invites,
         ]);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 }
