+++ app/Http/Livewire/NotificationSearch.php
@@ -14,11 +14,13 @@
 namespace App\Http\Livewire;
 
 use App\Models\User;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class NotificationSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public bool $bon_gifts = false;
@@ -139,17 +141,6 @@ class NotificationSearch extends Component
             ->orderBy('is_read')
             ->orderBy($this->sortField, $this->sortDirection)
             ->paginate($this->perPage);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
