+++ app/Http/Livewire/SubtitleSearch.php
@@ -16,11 +16,13 @@ namespace App\Http\Livewire;
 use App\Models\Subtitle;
 use App\Models\Torrent;
 use App\Models\User;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class SubtitleSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public int $perPage = 25;
@@ -75,17 +77,6 @@ class SubtitleSearch extends Component
             )
             ->orderBy($this->sortField, $this->sortDirection)
             ->paginate($this->perPage);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
