+++ app/Http/Livewire/UserTorrents.php
@@ -15,11 +15,13 @@ namespace App\Http\Livewire;
 
 use App\Models\History;
 use App\Models\User;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class UserTorrents extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public ?User $user = null;
@@ -177,16 +179,5 @@ class UserTorrents extends Component
         return view('livewire.user-torrents', [
             'histories' => $this->history,
         ]);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 }
