+++ app/Http/Livewire/UserWarnings.php
@@ -16,6 +16,7 @@ namespace App\Http\Livewire;
 use App\Models\PrivateMessage;
 use App\Models\User;
 use App\Models\Warning;
+use App\Traits\LivewireSort;
 use Illuminate\Support\Carbon;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -28,6 +29,8 @@ use Livewire\WithPagination;
  */
 class UserWarnings extends Component
 {
+    use LivewireSort;
+    use LivewireSort;
     use WithPagination;
 
     public User $user;
@@ -255,17 +258,6 @@ class UserWarnings extends Component
         Warning::withTrashed()->findOrFail($id)->restore();
 
         $this->dispatchBrowserEvent('success', ['type' => 'success', 'message' => 'Warning Was Successfully Restored']);
-    }
-
-    final public function sortBy(string $field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
