+++ app/Http/Livewire/WatchlistSearch.php
@@ -14,11 +14,13 @@
 namespace App\Http\Livewire;
 
 use App\Models\Watchlist;
+use App\Traits\LivewireSort;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class WatchlistSearch extends Component
 {
+    use LivewireSort;
     use WithPagination;
 
     public ?\Illuminate\Contracts\Auth\Authenticatable $user = null;
@@ -56,17 +58,6 @@ class WatchlistSearch extends Component
             ->when($this->search, fn ($query) => $query->where('message', 'LIKE', '%'.$this->search.'%'))
             ->orderBy($this->sortField, $this->sortDirection)
             ->paginate($this->perPage);
-    }
-
-    final public function sortBy($field): void
-    {
-        if ($this->sortField === $field) {
-            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
-        } else {
-            $this->sortDirection = 'asc';
-        }
-
-        $this->sortField = $field;
     }
 
     final public function render(): \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Contracts\Foundation\Application
